fixing display of lookbook in home and delete products relationship

+++ app/models/lookbook.rb
@@ -1,10 +1,10 @@
 class Lookbook < ActiveRecord::Base
   validates :name, :presence => true, :uniqueness => true
-	has_many :images, :dependent => :destroy
-	has_many :lookbooks_products, :dependent => :destroy
+  has_many :images, :dependent => :destroy
+  has_many :lookbooks_products, :dependent => :destroy
   has_many :products, :through => :lookbooks_products
 
-	accepts_nested_attributes_for :images, :reject_if => lambda{|p| p[:image].blank?}
+  accepts_nested_attributes_for :images, :reject_if => lambda{|p| p[:image].blank?}
 
   mount_uploader :thumb_image, ImageUploader
 
@@ -16,8 +16,8 @@ class Lookbook < ActiveRecord::Base
 
   def update_products
     
-    products.delete_all
-		selected_products = product_list.nil? ? [] : product_list.keys.collect{|id| Product.find_by_id(id)}
+    products.delete_all unless product_list.nil?
+    selected_products = product_list.nil? ? [] : product_list.keys.collect{|id| Product.find_by_id(id)}
     selected_products.each {|os| self.products << os}
     lb = self.lookbooks_products
     lb.each do |lb_product|
